requirements.txt
# Mentor Scoring AI - Dependencies
# All dependencies are CPU-compatible, no GPU required

# Web Framework
streamlit>=1.31.0

# Environment variables
python-dotenv>=1.0.0

# Speech-to-Text (OpenAI Whisper API - much faster than CPU-based)
openai>=1.0.0

# Content Evaluation (Groq API - much faster than local Ollama)
groq>=0.4.0

# Computer Vision (Posture Analysis)
# IMPORTANT: MediaPipe 0.10.9 is required for mp.solutions.pose API
mediapipe==0.10.9
opencv-python>=4.9.0

# TensorFlow (MediaPipe dependency - pinned for compatibility)
tensorflow==2.15.0

# Protobuf (pinned for MediaPipe/TensorFlow compatibility)
protobuf==4.23.4

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1  # Required by librosa for audio I/O

# Audio/Video Processing
ffmpeg-python>=0.2.0

# Numerical Computing
numpy>=1.24.0,<2.0.0  # MediaPipe compatibility

# Protobuf (pinned for MediaPipe/TensorFlow compatibility)
protobuf==4.25.3

# HTTP Requests (for Ollama)
requests>=2.31.0

# Utilities
tqdm>=4.66.0  # Progress bars for processing
